odb-helper: add write_object_process()

+++ external-odb.h
@@ -11,7 +11,7 @@ typedef int (*each_external_object_fn)(const unsigned char *sha1,
 				       unsigned long size,
 				       void *data);
 int external_odb_for_each_object(each_external_object_fn, void *);
-int external_odb_write_object(const void *buf, unsigned long len,
+int external_odb_write_object(const void *buf, size_t len,
 			      const char *type, unsigned char *sha1);
 
 #endif /* EXTERNAL_ODB_H */
